refactor: 迁移 Todo 查询新鲜度 helper
