+++ qq-maid-core/src/runtime/tools/todo/mod.rs
@@ -16,6 +16,7 @@ mod common;
 pub(crate) mod edit_patch;
 pub(crate) mod flow;
 pub(crate) mod format;
+mod freshness;
 mod json;
 pub(crate) mod ops;
 pub(crate) mod pending;
@@ -47,6 +48,7 @@ pub use create::CreateTodoTool;
 pub use delete::DeleteTodoTool;
 pub use edit::EditTodoTool;
 pub use edit_patch::TodoEditPatch;
+pub(crate) use freshness::valid_last_visible_todo_query;
 pub use get::GetTodoTool;
 pub use list::ListTodoTool;
 pub use merge::MergeTodoTool;
